+++ src/dsp.zig
@@ -30,6 +30,7 @@ pub const Clipper = @import("dsp/clipper.zig").Clipper;
 pub const clipper = @import("dsp/clipper.zig");
 pub const Gate = @import("dsp/gate.zig").Gate;
 pub const gate = @import("dsp/gate.zig");
+pub const amp = @import("dsp/amp.zig");
 pub const Saturator = @import("dsp/saturator.zig").Saturator;
 pub const Crusher = @import("dsp/crusher.zig").Crusher;
 pub const Chorus = @import("dsp/chorus.zig").Chorus;